Refactor : extract classes
