+++ RealWord.DB/Repositories/Interfaces/IArticleRebository.cs
