+++ RealWord.DB/Repositories/Interfaces/ILikeRepository.cs
