+++ RealWord.DB/Services/Interfaces/ICommentService.cs
