+++ RealWord.DB/Services/Interfaces/ILikeService.cs
